fix: move upload cover endpoint to project
